requirements.txt
airflow-provider-duckdb
polars
ipython
pyiceberg
pyarrow
